+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # HEAD
 
+## Bugfixes
+- Change `Shape` -> `AbstractShape`
+- Update documentation in `shapes.jl`
+- Correct typo `ray_interception` -> `ray_intersection`
+- Wrong operator in `ray_interception(pl::Plane, ray::Ray)` [#12](https://github.com/stevelonny/jujutracer/issues/7)
+
 # Version 0.2.0
 - Implement geometry operations [#1](https://github.com/stevelonny/jujutracer/pull/1)
 - Camera types [#2](https://github.com/stevelonny/jujutracer/pull/2)